Refactor entities package: Split schemas into separate files
- Split 436-line index.ts into 8 focused files by domain
- Each file now has comprehensive JSDoc documentation
- Maintained backward compatibility with re-exports
- Fixed circular dependency with DocumentSetSchema
- Updated turbo.json: renamed 'pipeline' to 'tasks' (v2.0 requirement)
- Removed orphaned test-reporter.ts from core-operations

Benefits:
- Better code organization and discoverability
- Easier navigation (files are 50-120 lines each)
- Reduced merge conflicts for team development
- Clear separation of concerns by domain

No breaking changes - all imports continue to work as before.
